License select & save rewrite finished
I've finished the selection of a license for the network settings, site
settings and rewrote it for the user profile & post/page interfaces. It
should be possible to select a license from all four possible
settings in WordPress from now on.

In the default mode (in which site admins may change license and users
may change their license), the network license is used as a default for
a site, the site default is used for a user profile and the user profile
license is used for the page or post.

My idea  is to use the same methodology (reversed & based on checking
the current settings) for displaying a license. So in the default mode,
it starts looking for license data from the post/page upward towards
the network or site license depending on if the plugin is used in a
network or not. It used the current settings to do so. In other words if
the post/page has a license, but the site admin has decided that the
site license should be used, the site license will be used instead of
the post/page license.

Possible future changes:

* Change the saved option name to _license so this is consistent over
  all the code. Now I'm using the 'license' name (without underscore)
  except for post/page post_meta, because without the underscore the
  name of the post_meta_key is used in the custom field metabox :( even
  though I'm using a serialized array...

* Add an option for users to make their default license the default
  across all their sites within one netwerk. See the global param:
  http://codex.wordpress.org/Function_Reference/update_user_option