First attempt to restore displaying the license
The code for getting the right license needs more testing, because I
have the feeling this is not the best way to select a license. Also need
to look at pages where there's more than one post (archives, search
results, home page etc) and how to deal with the differences in
licenses (especially considering the widget which implies one licens for
all).