Include locale in license chooser
Using the lang parameter in the call for the CC partner url we can
change the license chooser's interface language. The language is grabbed
from the get_locale() function in WordPress and should work even if the
locale is not supported by the license chooser since it default to en_US
automagically. Fixed #12