Fixed #16 Added warning at site/network settings
I've added a warning text at the site and network settings that changing
the settings here will relicense content. For instance if a site admin
decides that he/she does not want users to have the ability to change
the license of their content, the previously written content will
automatically be relicensed to the site's default license. We need to
have a look at this with feedback from people using the plugin to see if
this approach is the right one.