Changed default license to use 4.0 instead of 3.0
By default the plugin was using the old 3.0 version of the Creative
Commons BY-SA license. This has now been changed to use the new 4.0
version of the Creative Commons BY-SA license by default. NB: when
selecting a license by the user it was already using the 4.0 version of
the licenses.